refactor(identity): Channel's member NAME rendering resolves via src/identity
Row 23. useMembersWithPublicProfileFallback used to be the NAME source
for Channel's message rendering; it no longer is - the message header,
mention pills, the profile card, and mention-pill clicks all resolve
via src/identity now (earlier Phase D rows), keyed on address alone.
Verified caller-by-caller that none of them read this hook's output
anymore.

What's still load-bearing and had to keep working: userIcon/bio
(avatars - src/identity deliberately doesn't resolve these, see
MemberName's caller-supplied userIcon prop) and the raw
primaryUsername/globalDisplayName tiers useMentionInput.ts's
mention-autocomplete search matching reads directly (outside this
migration's scope). displayName is kept too, for one remaining live
consumer (the reply-heading preview text via quorum-shared's
replaceMentionsWithDisplayNames), but its public-profile fetch tier is
dropped - it now reads only the raw roster. Channel.tsx cannot route
displayName through src/identity itself: mapSenderToUser is built in
Channel's own function body, which runs before <IdentityScopeProvider>
(a descendant of what Channel returns, not an ancestor) - the same
constraint ChannelTypingIndicator already works around by being a
separate child component.

Renamed to useVisibleSenderProfileFallback and moved to its own file;
useMembersWithPublicProfileFallback.ts is deleted. Pinned the dropped
fetch tier with a test proven RED by re-adding it.
